backend/llm_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.0
redis==5.0.1
pytest==7.4.4
pytest-asyncio==0.21.1
